add method to fix data type
