+++ .gitignore
@@ -1,12 +1,23 @@
-# Atari 2600 (Stella) support for multiple assemblers
-# - DASM
-# - CC65
+# ignore everything in the root except the "wp-content" directory.
+!wp-content/
 
-# Assembled binaries and object directories
-obj/
-a.out
-*.bin
-*.a26
+# ignore everything in the "wp-content" directory, except:
+# "mu-plugins", "plugins", "themes" directory
+wp-content/*
+!wp-content/mu-plugins/
+!wp-content/plugins/
+!wp-content/themes/
 
-# Add in special Atari 7800-based binaries for good measure
-*.a78
+# ignore these plugins
+wp-content/plugins/hello.php
+
+# ignore specific themes
+wp-content/themes/twenty*/
+
+# ignore node dependency directories
+node_modules/
+
+# ignore log files and databases
+*.log
+*.sql
+*.sqlite
