+++ README.md
@@ -1,1 +1,51 @@
-# Enumerables
+# bubbleSort
+
+> This project shows sorting algorithms of integer arrays and string arrays
+
+![Screenshot](./big-sc.png)
+
+This is our 1st ruby project in microverse
+
+## Built With
+
+- Ruby
+
+## Live Demo
+
+https://replit.com/@vikitaotiz/Bubble-Sort-1#main.rb
+
+## Getting Started
+
+1. Clone the project
+   https://github.com/igorkol91/bubbleSort/tree/development
+
+2. Then open https://replit.com/languages/ruby
+
+Copy and paste the code in replit.com
+
+### Prerequisites
+
+Just ensure you have one or more web browsers
+
+## Authors
+
+:bust_in_silhouette: **Author1**
+
+- GitHub: [@igorkol91](https://github.com/igorkol91)
+- Twitter: [@Destro49536502]
+- LinkedIn: [LinkedIn](https://linkedin.com/in/linkedinhandle)
+
+:bust_in_silhouette: **Author2**
+
+- GitHub: [@vikitaotiz](https://github.com/vikitaotiz)
+- Twitter: [@victoro29641869]
+- LinkedIn: [LinkedIn](https://linkedin.com/in/linkedinhandle)
+
+## :handshake: Contributing
+
+Contributions, issues, and feature requests are welcome!
+Feel free to contribute
+
+## Show your support
+
+Give a ⭐️ if you like this project!
